includes the start of basic AI for blinky
This will need to be updated to check on each frame whether blinky
should be continuing to move forward or needs to turn either right or
left based on the Pacman position.

Currently, if the forward path is clear, Blinky will continue moving
forward regardless of the Pacman position. This is what needs to be
updated.