Pacman Die
Added functionality to reset Pacman’s position to default upon contact
with objects that I tagged as “Ghosty”.

I was unable to get the Orange or Blue ghosts to trigger this action,
possibly due to the presence of Player Controller components on them
rather than a collider. I decided not to mess with this to avoid
stepping on the toes of those programming the ghost movement.

I had to edit the DestroyPellet script so that the ghosts were not
destroyed upon contact with the player.

I added a Canvas to display when the life counter reaches zero and used
the Return key to retrigger the initial scene.